requirements.txt
aiogram
netschoolapi
openpyxl
python-dotenv